refactor game component location
